+++ R/time_mutate.R
@@ -150,7 +150,7 @@ time_mutate <- function(data, time = NULL, ..., time_by = NULL,
       out <- dplyr::dplyr_col_modify(out,
                                      add_names(
                                        list(
-                                         time_interval(time_agg, time_int_end)
+                                         time_interval2(time_agg, time_int_end)
                                        ), int_nm
                                      )
       )